Renaming the packages from 'ops' to their respective command names
